requirements.txt
botbuilder-integration-aiohttp
openai==1.42.0
